Almost complete the data packet system.
The data packet system is now working for all *_INTERFACE's and
NORMAL_FORCE. Remenants for the old dumpData remains. In order
to have a cohesive way of writing data packets, all classes
that do write data packets are derived from the simple class dumpable, even
when this results in multiple inheritance. The getDataPackets-chain is not
main->sidePotentialloading->lattice & springFriction