Began writing script for analyzing data.
The script is simply a rewrite of the old scripts, but somewhat
more organized. There are two classes - DataReader and Analyzer.
The DataReader simply provides an interface to read the binary
files and to the necessary operations on them, while the Analyzer
uses DataReader to read all the files and plot them accordingly