Added Parameters class
The parameters class reads the entire configuration file
line by line, and cuts each line into its substrings (i.e "nx 5 lattice
length" -> ["nx", "5", "lattice", "length"]).
The first substring is matched to a parameter name (i.e "nx") while the
second is used as the parameter's value (5). The rest of the substrings
are ignored. The parameter is then converted into float. If the value should
be an int, the converted value is then cast into an int.

Once a parameter is matched, the corresponding parameter is set. This parameter
is always named by m_<parametername>. A boolean is set to ensure that the
parameter is set. If not all parameters are set once the file is read, a
runetimeerror is thrown.