Fixed some errors and added friction type to parameters
The Van Gogh rotation was caused by a line of test code pertubating the rotation
of each node enough to cause an eternal rotation.

The snapshots is now working. Before, the same filehandle was used for both
standard output and snapshots, causing the os to get mad and refuse to open the
files.

The type of friction system (toploading, sideloading, cantilever, etc) is now
set in the parameters with the keyword "frictionsystem". The case of the value
is irrelevant, as it is transformed to lower case in the program.

The version is now 0.91 since the parameters are incompatible with earlier
versions.