Extended graphical analyzer
The window providing more detailed information is now a QtCanvas, allowing
buttons to be added. A 'view lattice' button is added, causing the xyz-file in
output/info/lattice.xyz to be rendered into output/info/lattice.png if
necessary, and showing the result in a new figure.

The ticks on x- and y-axis are now discrete if the values they represent are
discrete.